need debug pipe

+++ Sources/mandatory_function/utils_mandatory.c
@@ -26,25 +26,23 @@ void	exe_cmd(t_data *data)
 
 void	exe_pipe(t_data *data)
 {
-	int	status;
-
 	while (data->cmd)
 	{
-		status = 0;
-		// if (!init_pipe(data))
-		// 	return ;
+		if (pipe(data->cmd->pipefd) == -1)
+			return ;
 		data->cmd->pid = fork();
 		if (data->cmd->pid == -1)
 			return (ft_strerror(data->cmd->cmd, NULL, strerror(errno)));
 		else if (data->cmd->pid > 0)
 		{
-			waitpid(data->cmd->pid, &status, 0);
+			waitpid(data->cmd->pid, &data->cmd->status, 0);
 			kill(data->cmd->pid, SIGTERM);
 			data->cmd = data->cmd->next;
 		}
 		else
 			child(data);
 	}
+	close_all_pid(data);
 }
 
 int		is_builting_cmd(t_data *data)
